### 2024-08-15 #7106 @editorialbot
[PRE REVIEW]: constrained_manipulability: A ROS 2 library to Compute and Visualize Constrained Capacities for Robotic Manipulators
**Submitting author:** @mazrk7 (<a href="http://orcid.org/0000-0002-7672-940X">Mark Zolotas</a>)
**Repository:** https://github.com/philip-long/constrained_manipulability
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
