### 2024-08-23 #7106 @crvernon
:wave: @Nikoleta-v3 are you able to take this one on as editor?  If so, there is also a companion submission that the authors submitted at the same time that uses this package as a dependency (see https://github.com/openjournals/joss-reviews/issues/7107#issuecomment-2291721853).  Would you also consider taking that one to keep the process cleaner?
