### 2024-09-09 #7106 @crvernon
:wave: @mazrk7 - I think I'll take these submissions myself.  Please do me a favor up front and lower the word count of the paper to about 1000 words. Most detail can usually be moved to the package documentation.

Also, do you have recommendations for reviewers that you would like me to check out?  Thanks!
