+++ paper/paper.md
@@ -43,7 +43,7 @@ The main program of the `constrained_manipulability` package reads a kinematic c
 
 # Statement of Need
 
-Generating complex constraint geometries in an efficient and effective manner is a significant challenge for real-time robot manipulation. For applications, like motion planning or remote teleoperation, being able to represent the space of allowable motion around a robot's end-effector is a vital capability. Given such a representation of the obstacle-free volume around the robot's end-effector, a collision-free IK solution can be obtained to maximize the free space of a robot moving in a constrained environment. Manipulability polytopes represent manipulability capacities, e.g., for the robot to transmit velocities and forces from the configuration space to the task space. The benefit of using convex polytopes is that they capture exact velocity bounds, rather than the approximation provided by other typical choices for performance indicators, such as ellipsoids [@Yoshikawa1984Analysis]. While the aforementionned ellipsoids have been the dominant paradigm due to historic computational constraints, more efficicent polytope generation methods [@Skuric2023; @Sagar2023] coupled with more computation availability has led to an increased usage.   
+Generating complex constraint geometries efficiently is a significant challenge for real-time robot manipulation. For applications, like motion planning or remote teleoperation, being able to represent the space of obstacle-free allowable motion around an end-effector is vital. Given such a representation a collision-free IK solution can be obtained to maximize the free space of a robot moving in a constrained environment. Manipulability polytopes represent manipulability capacities, e.g., for the robot to transmit velocities and forces from the configuration space to the task space. The benefit of using convex polytopes is that they capture exact velocity bounds, rather than the approximation provided by for example ellipsoids [@Yoshikawa1984Analysis]. While the aforementionned ellipsoids have been the dominant paradigm due to historic computational constraints, more efficicent polytope generation methods [@Skuric2023; @Sagar2023] coupled with more computation availability has led to an increased usage.   
 Furthermore, since a polytope is defined by a set of inequality constraints, additional constraints can be easily incorporated into existing polytopes, e.g., mobile robot toppling constraints [@rasheed2018tension], friction cones [@caron2016zmp], and maximum danger values. 
 
 The `constrained_manipulability` ROS 2 package aims to fulfil this need for constrained robot capacity calculation by supplying the robotics community with a fast C++ implementation that computes various types of polytopes. These geometrical constructs can also be visualized for workspace analysis or to guide an operator through the Cartesian motions available due to joint limits, kinematic constraints, and obstacles in the workspace, as illustrated in \autoref{fig:traj_planning}. The utility of these visualizations has proven advantageous in remote teleoperation scenarios involving virtual reality headsets [@Zolotas2021Motion]. Moreover, the package provides an interface to the `robot_collision_checking` package and a set of convenience functions for checking the distance/collision between a robot manipulator and the environment.
@@ -52,7 +52,7 @@ The `constrained_manipulability` ROS 2 package aims to fulfil this need for cons
 
 # Manipulability Polytopes Calculation
 
-A polytope, $\mathcal{P}$, can be represented in two ways: as the convex hull of its vertex set, known as the $\mathcal{V}$-representation and denoted as $\mathcal{P}^{V}$, or as a volume bounded by a finite number of half-spaces, known as the $\mathcal{H}$-representation and denoted as $\mathcal{P}^{H}$. These are written respectively as:
[…]
