+++ paper/paper.md
@@ -50,19 +50,8 @@ The `constrained_manipulability` ROS 2 package aims to fulfil this need for cons
 
 ![Collision-free path for a robot manipulator visualized as a trajectory of polytopes generated by maximizing the volume of allowable motion around the robot's end-effector.\label{fig:traj_planning}](figures/trajplanning.png)
 
-# Manipulability Polytopes Calculation
-
-A polytope, $\mathcal{P}$, can be represented as the convex hull of its vertex set, or as a volume bounded by a finite number of half-spaces,known as the $\mathcal{V}$-representation and $\mathcal{H}$-representation and denoted as $\mathcal{P}^{V}$ and $\mathcal{P}^{H}$, respectively.
-\begin{equation} \label{eq:poly_spaces}
-\begin{split}
-\mathcal{P}^{V} &= \{\mathbf{x}: \mathbf{x}=\sum\limits_{i=1}^{n} \alpha_{i}\mathbf{y}_{i}\Bigg| \alpha_{i}\geq0, \sum\limits_{i=1}^{n}\alpha_{i}=1\}, \\ \mathcal{P}^{H} &= \mathbf{A}\mathbf{x} \leq \mathbf{b},
-\end{split}
-\end{equation}
-where $\mathbf{y}_{i}$ denotes the $i^{th}$ element of the vertex set, $\mathbf{x}$ is any point inside $\mathcal{P}$, $\mathbf{A}$ contains the half-spaces' normals, and $\mathbf{b}$ is the shifted distance from the origin along the normal. Converting between the $\mathcal{V}$ and $\mathcal{H}$ representations can be carried out in several ways, however the `constrained_manipulability` package uses the double description method [@Fukuda1996Double]. 
-
-Below is a description of how constrained motion polytopes are constructed in `constrained_manipulability` for a given manipulator model.
-
 ## Manipulator Model
+A polytope, $\mathcal{P}$, can be represented as the convex hull of its vertex set, or as a volume bounded by a finite number of half-spaces,known as the $\mathcal{V}$-representation and $\mathcal{H}$-representation and denoted as $\mathcal{P}^{V}$ and $\mathcal{P}^{H}$, respectively. Converting between the $\mathcal{V}$ and $\mathcal{H}$ representations can be carried out in several ways, however the `constrained_manipulability` package uses the double description method [@Fukuda1996Double]. 
 
 Consider a serial manipulator with $n$ degrees-of-freedom operating in 6-dimensional space. The end-effector pose is represented by a vector $\mathbf{x}_n \in \mathbb{R}^6$, consisting of the position $\mathbf{x}^p_n \in \mathbb{R}^3$ and unit quaternion orientation $\mathbf{x}^q_n \in \mathbb{R}^3$, all of which can be derived from the forward kinematic chain $\mathit{fk}$:
 \begin{equation} \label{eq:fk}
