+++ paper/paper.md
@@ -50,40 +50,41 @@ The `constrained_manipulability` ROS 2 package aims to fulfil this need for cons
 
 ![Collision-free path for a robot manipulator visualized as a trajectory of polytopes generated by maximizing the volume of allowable motion around the robot's end-effector.\label{fig:traj_planning}](figures/trajplanning.png)
 
-## Manipulator Model
+## Polytope and Manipulator Model
 A polytope, $\mathcal{P}$, can be represented as the convex hull of its vertex set, or as a volume bounded by a finite number of half-spaces,known as the $\mathcal{V}$-representation and $\mathcal{H}$-representation and denoted as $\mathcal{P}^{V}$ and $\mathcal{P}^{H}$, respectively. Converting between the $\mathcal{V}$ and $\mathcal{H}$ representations can be carried out in several ways, however the `constrained_manipulability` package uses the double description method [@Fukuda1996Double]. 
 
-Consider a serial manipulator with $n$ degrees-of-freedom operating in 6-dimensional space. The end-effector pose is represented by a vector $\mathbf{x}_n \in \mathbb{R}^6$, consisting of the position $\mathbf{x}^p_n \in \mathbb{R}^3$ and unit quaternion orientation $\mathbf{x}^q_n \in \mathbb{R}^3$, all of which can be derived from the forward kinematic chain $\mathit{fk}$:
+Consider a $n$ degrees-of-freedom serial manipulator, whose end-effector pose is represented by a vector $\mathbf{x}_n \in \mathbb{R}^6$, consisting of the position $\mathbf{x}^p_n \in \mathbb{R}^3$ and unit quaternion orientation $\mathbf{x}^q_n \in \mathbb{R}^3$, all of which can be derived from the forward kinematic chain $\mathit{fk}$:
 \begin{equation} \label{eq:fk}
     \mathbf{x}_n = \mathit{fk}(\mathbf{q}),
 \end{equation}
-where $\mathbf{q}=[q_1, ..., q_n]$ are the joint configuration variables. The twist at the end-effector's $n^{th}$ frame, $\boldsymbol{\nu}_n$, can then be obtained from the differential kinematic model:
+where $\mathbf{q}=[q_1, ..., q_n]$ are the joint configuration variables. The twist at the end-effector's $n^{th}$ frame, $\boldsymbol{\nu}_n$, can then be obtained as:
 \begin{equation} \label{eq:dk}
     \boldsymbol{\nu}_n = \begin{bmatrix} \mathbf{v} \\ \boldsymbol{\omega} \end{bmatrix} = \mathbf{J}_n \dot{\mathbf{q}},
 \end{equation}
-with $\mathbf{v}$ and $\boldsymbol{\omega}$ denoting translational and angular velocities, respectively. $\mathbf{J}_n$ is the 6$\times n$ Jacobian matrix defined at the end-effector frame $n$ and $\dot{\mathbf{q}} = [\dot{q}_1, ..., \dot{q}_n]^\intercal$ is the joint velocity vector. 
+with $\mathbf{v}$ and $\boldsymbol{\omega}$ denoting translational and angular velocities, respectively. $\mathbf{J}_n$ is the 6$\times n$ Jacobian matrix defined at $n$ and $\dot{\mathbf{q}} = [\dot{q}_1, ..., \dot{q}_n]^\intercal$ is the joint velocity. 
 
 ## Allowable Motions
 
[…]
