+++ paper/paper.md
@@ -35,15 +35,15 @@ bibliography: paper.bib
 
 # Summary
 
-This paper presents `constrained_manipulability`, a C++ library to compute and visualize a robot manipulator's constrained motion capacities during manipulation tasks. Manipulability polytopes provide a geometrical tool to evaluate the volume of free space surrounding the robotic arm when considering both environmental constraints and intrinsic robot constraints, such as environmental collisions and a manipulator's joint limits, respectively. Moreover, the polytopes defined by these convex constraints represent feasible configurations for the whole robot, that can be utilized in inverse kinematics (IK) optimization algorithms to produce collision-free motion trajectories.
+This paper presents `constrained_manipulability`, a C++ library to compute and visualize a robot manipulator's constrained motion capacities. Manipulability polytopes provide a geometrical tool to evaluate the volume of free space surrounding the robotic arm when considering both environmental and intrinsic robot constraints, such as collisions and joint limits, respectively. Moreover, the polytopes defined by these convex constraints represent feasible configurations for the whole robot, that can be utilized in inverse kinematics (IK) optimization algorithms to produce collision-free motion trajectories.
 
 The library is encapsulated in a Robot Operating System (ROS) package [@Quigley2009ROS]. We include ROS 1 and ROS 2 [@Macenski2022ROS2] implementations of the core C++ library. The `constrained_manipulability` package also heavily depends on our `robot_collision_checking` package, a ROS interface for collision checking via the Flexible Collision Library (FCL) [@Pan2012FCL].
 
-The main program of the `constrained_manipulability` package reads a kinematic chain from a robot manipulator's Unified Robot Description Format (URDF), from a user defined root of the robot (e.g., the robot's base frame) to a tip (e.g., the end-effector or tool). Joint position and velocity limits are also read into the program using this URDF input, while a collision world is maintained to express environmental constraints on the robot's motion. A variety of collision objects can be added or removed using common ROS message types, such as Octomaps [@Hornung2013Octomap], mesh files or solid primitives. These joint limit and obstacle constraints are then stacked together to define different polytopes, such as the allowable motion polytope [@Long2019Optimization] and the constrained velocity polytope [@Long2018Humanoids].
+The main program of the `constrained_manipulability` package reads a kinematic chain from a robot manipulator's Unified Robot Description Format (URDF), from a user defined root of the robot (e.g., the robot's base frame) to a tip (e.g., the end-effector or tool). Joint position and velocity limits are also read from the URDF, while a collision world is maintained to express environmental constraints on the robot's motion. A variety of collision objects can be added or removed using common ROS message types, such as Octomaps [@Hornung2013Octomap], mesh files or solid primitives. These joint limit and obstacle constraints are then stacked together to define different polytopes, such as the allowable motion polytope [@Long2019Optimization] and the constrained velocity polytope [@Long2018Humanoids].
 
 # Statement of Need
 
[…]
