+++ paper/paper.bib
@@ -8,6 +8,14 @@
     journal = {ICRA Workshop on Open Source Software}
 }
 
+@inproceedings{kokkinis1989kinetostatic,
+  title={Kinetostatic performance limits of cooperating robot manipulators using force-velocity polytopes},
+  author={Kokkinis, T and Paden, B},
+  booktitle={Proceedings of the ASME Winter Annual Meeting},
+  pages={151--155},
+  year={1989}
+}
+
 @article{Macenski2022ROS2,
     author = {Steven Macenski  and Tully Foote  and Brian Gerkey  and Chris Lalancette  and William Woodall},
     title = {Robot Operating System 2: Design, architecture, and uses in the wild},
