+++ paper/paper.md
@@ -53,21 +53,9 @@ The `constrained_manipulability` ROS 2 package aims to fulfil this need for cons
 ## Polytope and Manipulator Model
 A polytope, $\mathcal{P}$, can be represented as the convex hull of its vertex set, or as a volume bounded by a finite number of half-spaces,known as the $\mathcal{V}$-representation and $\mathcal{H}$-representation and denoted as $\mathcal{P}^{V}$ and $\mathcal{P}^{H}$, respectively. Converting between the $\mathcal{V}$ and $\mathcal{H}$ representations can be carried out in several ways, however the `constrained_manipulability` package uses the double description method [@Fukuda1996Double]. 
 
-Consider a $n$ degrees-of-freedom serial manipulator, whose end-effector pose is represented by a vector $\mathbf{x}_n \in \mathbb{R}^6$, which can be obtain from the forward kinematic chain $\mathit{fk}$ as follows:
-\begin{equation} \label{eq:fk}
-    \mathbf{x}_n = \mathit{fk}(\mathbf{q}),
-\end{equation}
-where $\mathbf{q}=[q_1, ..., q_n]$ are the joint configuration variables. The twist at the end-effector's $n^{th}$ frame, $\boldsymbol{\nu}_n$, can then be obtained as:
-\begin{equation} \label{eq:dk}
-    \boldsymbol{\nu}_n = \begin{bmatrix} \mathbf{v} \\ \boldsymbol{\omega} \end{bmatrix} = \mathbf{J}_n \dot{\mathbf{q}},
-\end{equation}
-with $\mathbf{v}$ and $\boldsymbol{\omega}$ denoting translational and angular velocities, respectively. $\mathbf{J}_n$ is the 6$\times n$ Jacobian matrix defined at $n$ and $\dot{\mathbf{q}} = [\dot{q}_1, ..., \dot{q}_n]^\intercal$ is the joint velocity. 
-
 The `constrained_manipulability` provides two types of polytope representation, Constrained Motion Polytopes which modify the classical manipulability velocity polytope [@kokkinis1989kinetostatic] to include joint limits and capacities reductions due to nearby obstacles and the Allowable Motion Polytope a linearisation of the above to generate a measure of free space or allowable motions in which a robot can move while satisfying all constraints. 
 
-## Constrained Motion Polytopes
-The velocity capacity can be obtained by first constructing the joint velocity space polytope in the $\mathcal{H}$-representation according to its velocity limits:
-\begin{equation} \label{eq:qh_repr}
+In both cases, the polytope is contructed as follows. First a system of linear inequalities is constructed in joint space, for instance based on the joint velocities limits: \begin{equation} \label{eq:qh_repr}
 \begin{split} 
 \mathcal{Q}^{H}=
 \left[\begin{array}{ccc} 
@@ -78,18 +66,14 @@ The velocity capacity can be obtained by first constructing the joint velocity s
 \end{array} \right], 
 \end{split} 
 \end{equation}
-where $\dot{\mathbf{q}}^{max}$ and $\dot{\mathbf{q}}^{min}$ are the robot's maximum and minimum joint velocities.
-
-Using the double description method, an equivalent polytope in the $\mathcal{V}$-representation (i.e., defined by its vertices) is written as:
[…]
