### 2024-09-10 a861c81
Reducing equation content and derivations
