### 2025-01-13 philip-long/constrained_manipulability#5 @JHartzer
JOSS: Add community guidelines
Please add clear guidelines for third parties wishing to
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

Some put this information in the main-level `README.md` file while others make a specific `CONTRIBUTING.md` file.
