### 2025-01-13 philip-long/constrained_manipulability#6 @JHartzer
JOSS: Unit Testing
There doesn't appear to be any way to test your code in a systematic way (i.e. with unit tests). At a minimum, I would like to see manual testing of the core functionality of the package. Automated testing would be preferred, but code coverage is a higher priority to ensure correct code behavior of the `ConstrainedManipulability` class.
