### 2025-01-13 philip-long/constrained_manipulability#7 @JHartzer
JOSS: Dependency on ur_description
Consider adding `<exec_depend>ur_description</exec_depend>` to your `package.xml` such that your examples can be run out of the box with rosdep-installed dependencies. Currently there are quite a few dependencies that have to be manually installed to get up and running, and this would be one fewer step to getting a first example running successfully.
