### 2025-01-13 #7481 @JHartzer
Overall I think the paper is well written and clearly explains the contributions of the package. It is not immediately clear that no other packages can plot constrained manipulator spaces, so that might be worth highlighting more. I've summarized my review below:

## Paper Issue:
- [ ] It looks like there is a dead reference to `\autoref{eq:joint_space_h_repr}` that needs to be fixed

## Software Issues:
- [ ] https://github.com/philip-long/constrained_manipulability/issues/5
- [ ] https://github.com/philip-long/constrained_manipulability/issues/6
- [ ] https://github.com/philip-long/constrained_manipulability/issues/7

## Suggestions:
- "The main program of the constrained_manipulability package..." is a bit of a run-on. Consider rewording to "..package reads a Unified Robot Description Format (URDF) kinematic chain from a user..."
- Is the paragraph break before "Furthermore, since a polytope is defined..." intentional?
- Consider adding a colon and "and" to "For the Constrained Motion Polytopes, the following..."
