### 2025-01-18 38a48c8
Fixing broken equation reference
