+++ paper/paper.md
@@ -86,7 +86,7 @@ where $\dot{\mathbf{q}}^{v}_{i}$ denotes the $i^{th}$ vertex of the polytope $\m
 \end{equation}
 with $\boldsymbol{\nu}^{v}_{j} = \mathbf{J}_{n} \dot{\mathbf{q}}^{v}_{j}$ and $p \leq q$. The convexity of a polytope is preserved under affine transformation, thus a bounded volume of $\mathcal{MP}$ that represents the system's manipulability can easily be obtained to serve as an exact indicator of robot performance. 
 
-For the Constrained Motion Polytopes, the following constraints are supported, joint velocity limits, joint velocity damping due to positional joint limits, joint velocity damping due obstacle proximity to the kinematic chain. For the allowable motion polytope the following constraints are supported, positional joint limits, positional limits due obstacle proximity to the kinematic chain, linearization limits. Increasing the values of the linearisation expands the free space virtual fixture at a cost of reduced fidelity. Our implementation uses a scalar linearization limit for all joints that can also be altered at run time e.g. could be increased to enlarge the solution  or shrunk  to guide the user towards a defined goal configuration, as described in [@Zolotas2021Motion]. Finally, in our implementation, we select at every instant the point along each link nearest to all environmental obstacles. Hence, the system considers the set of instantaneous collision-free joint motion for each link on the robot's kinematic chain, considering all obstacles $\mathcal{O}$. 
+For the Constrained Motion Polytopes, the following constraints are supported, joint velocity limits, joint velocity damping due to positional joint limits, joint velocity damping due obstacle proximity to the kinematic chain. For the allowable motion polytope the following constraints are supported, positional joint limits, positional limits due obstacle proximity to the kinematic chain, linearization limits. Increasing the values of the linearisation expands the free space virtual fixture at a cost of reduced fidelity. Our implementation uses a scalar linearization limit for all joints that can also be altered at run time, e.g., could be increased to enlarge the solution or shrunk to guide the user towards a defined goal configuration, as described in [@Zolotas2021Motion]. Finally, in our implementation, we select at every instant the point along each link nearest to all environmental obstacles. Hence, the system considers the set of instantaneous collision-free joint motion for each link on the robot's kinematic chain, considering all obstacles $\mathcal{O}$. 
 
 ## Block Diagram
 
@@ -101,7 +101,7 @@ Polytope getConstrainedAllowableMotionPolytope(
   const std::vector<double>& color_pts,
   const std::vector<double>& color_line);
 ```
-where `AHrep` and `bHrep` represent the joint space polytope constraints defined in \autoref{eq:joint_space_h_repr}. 
+where `AHrep` and `bHrep` represent the joint space polytope constraints. 
 
 A diagram summarizing the `constrained_manipulability` package architecture and ROS 2 communication flow is displayed in \autoref{fig:block_diagram}.
 
