### 2025-01-18 8f1b61d
Minor touch-up of paper
