### 2025-01-18 765b44b
Adding dockerfile and updating readme
