### 2025-01-18 philip-long/constrained_manipulability#5 @mazrk7
Hi @JHartzer, thank you for the suggestion! Guidelines have been added to the repository.
