### 2025-01-18 philip-long/constrained_manipulability#7 @mazrk7
Thank you @JHartzer, the repository indeed required additional dependencies and clearer instructions for installation. A Dockerfile has since been added, with refinements made to the `README` in the hopes of making setup easier for end-users.
