### 2025-01-19 4e2b5b7
Minor changes to dockerfile
