### 2025-01-19 adf910e
Removed CI detail
