### 2025-01-19 philip-long/constrained_manipulability#6 @mazrk7
Hi @JHartzer, unit tests and CI have been added to the repository with the latest commits (see fbfd106 for `tests` additions). Please let me know if anything else is required. Thank you!
