+++ paper/paper.md
@@ -29,7 +29,7 @@ affiliations:
    index: 3
  - name: Amazon Robotics, USA
    index: 4
-date: 14 August 2024
+date: 20 January 2025
 bibliography: paper.bib
 ---
 
@@ -41,11 +41,15 @@ The library is encapsulated in a Robot Operating System (ROS) package [@Quigley2
 
 The main program of the `constrained_manipulability` package reads a Unified Robot Description Format (URDF) kinematic chain from a user-defined root of the robot (e.g., the robot's base frame) to a tip (e.g., the end-effector or tool). Joint position and velocity limits are also read from the URDF, while a collision world is maintained to express environmental constraints on the robot's motion. A variety of collision objects can be added or removed using common ROS message types, such as OctoMaps [@Hornung2013Octomap], mesh files, or solid primitives. These joint limit and obstacle constraints are then stacked together to define different polytopes, such as the allowable motion polytope [@Long2019Optimization] and the constrained velocity polytope [@Long2018Humanoids].
 
+Polytopes computed by `constrained_manipulability` are published as a `visualization_msgs::msg::MarkerArray` ROS message. A `visualization_msgs::msg::Marker::TRIANGLE_LIST` represents the polytope facets and a `visualization_msgs::msg::Marker::SPHERE_LIST` represents the vertices. As a result, the manipulability polytopes  calculated by our package for a given robot manipulator can also be visualized using standard RViz [@kam2015rviz] tools. Additionally, the generated polytopes are published in vertex and hyperplane form, which is easily interpretable by third-party libraries.
+
 # Statement of Need
 
 Generating complex constraint geometries efficiently is a significant challenge for real-time robot manipulation. For applications, like motion planning or remote teleoperation, being able to represent the space of obstacle-free allowable motion around an end-effector is vital. Given such a representation a collision-free IK solution can be obtained to maximize the free space of a robot moving in a constrained environment. Manipulability polytopes represent manipulability capacities, e.g., the capacity for the robot to transmit velocities and forces from the configuration space to the task space. The benefit of using convex polytopes is that they capture exact velocity bounds, rather than the approximation provided by for example ellipsoids [@Yoshikawa1984Analysis]. While the aforementionned ellipsoids have been the dominant paradigm due to historic computational constraints, more efficient polytope generation methods [@Skuric2023; @Sagar2023] coupled with more computation availability has led to an increased usage. Furthermore, since a polytope is defined by a set of inequality constraints, additional constraints can be easily incorporated into existing polytopes, e.g., mobile robot toppling constraints [@rasheed2018tension], friction cones [@caron2016zmp], and maximum danger values. 
 
 The `constrained_manipulability` ROS 2 package aims to fulfil this need for constrained robot capacity calculation by supplying the robotics community with a fast C++ implementation that computes various types of polytopes. These geometrical constructs can also be visualized for workspace analysis or to guide an operator through the Cartesian motions available due to joint limits, kinematic constraints, and obstacles in the workspace, as illustrated in \autoref{fig:traj_planning}. The utility of these visualizations has proven advantageous in remote teleoperation scenarios involving virtual reality headsets [@Zolotas2021Motion]. Moreover, the package interfaces with the `robot_collision_checking` package to perform distance and collision checking between a robot manipulator and the environment.
+
[…]
