+++ paper/paper.bib
@@ -137,3 +137,12 @@
     organization={Springer}
 }
 
+@article{kam2015rviz,
+  title={Rviz: a toolkit for real domain data visualization},
+  author={Kam, Hyeong Ryeol and Lee, Sung-Ho and Park, Taejung and Kim, Chang-Hun},
+  journal={Telecommunication Systems},
+  volume={60},
+  pages={337--345},
+  year={2015},
+  publisher={Springer}
+}
