### 2025-01-21 philip-long/constrained_manipulability#5 @JHartzer
Excellent! Thanks Mark
