### 2025-01-21 philip-long/constrained_manipulability#7 @JHartzer
Awesome!
