### 2025-01-21 philip-long/constrained_manipulability#6 @JHartzer
Excellent work, Mark! I was able to pull the changes and verify the test cases
