### 2025-01-21 #7481 @JHartzer
Thank you Mark for the quick responses to my feedback! You've addressed all concerns and issues I had so I'm content calling my review complete. One minor thing, though: the "visualization_msgs::msg::MarkerArray" in the above paper proof is hanging off the end of a line because it can't break. Might be worth rearranging some words or including a break.
