### 2025-01-28 3f27cd1
Adding COI statement and new paper citation
