### 2025-01-28 f463fbf
Adding missing bib reference
