+++ paper/paper.bib
@@ -146,3 +146,16 @@
   year={2015},
   publisher={Springer}
 }
+
+@article{Zolotas2025ROSFCL, 
+    doi = {10.21105/joss.07473}, 
+    url = {https://doi.org/10.21105/joss.07473}, 
+    year = {2025}, 
+    publisher = {The Open Journal}, 
+    volume = {10}, 
+    number = {105}, 
+    pages = {7473}, 
+    author = {Mark Zolotas and Philip Long and Taskin Padir}, 
+    title = {robot_collision_checking: A Lightweight ROS 2 Interface to FCL (Flexible Collision Library)}, 
+    journal = {Journal of Open Source Software} 
+}
