### 2025-02-22 b04fc28
Handling missing collision geometries along KDL chain for polytope plotting
