### 2025-02-22 cc76db9
README instructions for GUI display via Docker image
