### 2025-02-22 9e7c325
Updating paper with polytope specifics from prior work and future work section
