+++ paper/paper.bib
@@ -159,3 +159,16 @@
     title = {robot_collision_checking: A Lightweight ROS 2 Interface to FCL (Flexible Collision Library)}, 
     journal = {Journal of Open Source Software} 
 }
+
+@inproceedings{Carpentier2021Proximal,
+    TITLE = {{Proximal and Sparse Resolution of Constrained Dynamic Equations}},
+    AUTHOR = {Carpentier, Justin and Budhiraja, Rohan and Mansard, Nicolas},
+    URL = {https://hal.inria.fr/hal-03271811},
+    BOOKTITLE = {{Robotics: Science and Systems 2021}},
+    ADDRESS = {Austin / Virtual, United States},
+    YEAR = {2021},
+    MONTH = Jul,
+    PDF = {https://hal.inria.fr/hal-03271811/file/rss-proximal-and-sparse.pdf},
+    HAL_ID = {hal-03271811},
+    HAL_VERSION = {v1},
+}
