### 2025-03-15 #7481 @crvernon
:wave: @mazrk7 - Next is just setting up the archive for your new release.

We want to make sure the archive has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.

So here is what we have left to do on that front:

- [ ] Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service. Please ensure that the software archive uses the same license as the license you have posted on GitHub.

- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list also matches your paper exactly).

- [ ] Please respond with the DOI of the archived version here as well as the version of the software you released.

Once you adjust the references, I'll review your paper as well.
