+++ paper/paper.bib
@@ -63,6 +63,7 @@
     booktitle={International Conference on Robotics and Automation},
     pages={204--210},
     year={2019},
+    doi={10.1109/ICRA.2019.8794071},
 }
 
 @inproceedings{Long2018Humanoids,
@@ -71,6 +72,7 @@
 	booktitle={IEEE-RAS International Conference on Humanoid Robots},
     pages={497--502},
 	year={2018},
+    doi={10.1109/HUMANOIDS.2018.8624962},
 }
 
 @inproceedings{Fukuda1996Double,
@@ -80,7 +82,8 @@
     volume={1120},
     pages={91},
     year={1996},
-    organization={Springer Science \& Business Media}
+    organization={Springer Science \& Business Media},
+    doi={10.1007/3-540-61576-8_77},
 }
 
 @article{Hornung2013Octomap,
@@ -90,7 +93,8 @@
     volume={34},
     pages={189--206},
     year={2013},
-    publisher={Springer}
+    publisher={Springer},
+    doi = {10.1007/s10514-012-9321-0},
 }
 
 @article{Skuric2023, 
@@ -125,7 +129,8 @@
     number={1},
     pages={67--80},
     year={2016},
-    publisher={IEEE}
+    publisher={IEEE},
+    doi={10.1109/tro.2016.2623338},
 }
 
 @inproceedings{rasheed2018tension,
@@ -134,7 +139,8 @@
     booktitle={Cable-driven parallel robots: proceedings of the third international conference on cable-driven parallel robots},
     pages={268--279},
     year={2018},
-    organization={Springer}
+    organization={Springer},
+    doi={10.1007/978-3-319-61431-1_23},
 }
 
 @article{kam2015rviz,
@@ -144,7 +150,8 @@
   volume={60},
   pages={337--345},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11235-015-0034-5},
 }
 
 @article{Zolotas2025ROSFCL, 
@@ -171,4 +178,5 @@
     PDF = {https://hal.inria.fr/hal-03271811/file/rss-proximal-and-sparse.pdf},
     HAL_ID = {hal-03271811},
     HAL_VERSION = {v1},
+    doi={10.15607/rss.2021.xvii.017},
 }
