### 2025-03-16 #7481 @mazrk7
Hi @crvernon! Thanks for sending the next steps over.

I've adjusted the references to include the missing DOIs.

The DOI of the archived version on Zenodo is: [10.5281/zenodo.15033530](https://doi.org/10.5281/zenodo.15033530)

And the version number released is: v1.0

Please let me know if anything else is required of me.
