### 2025-04-11 #7481 @crvernon
👋  @mazrk7 - we are almost there!

Please address the following in your paper:

- [ ] LINES 30-31:  Your code formatting of "...published as a `visualization_msgs::msg::MarkerArray` ROS..." is causing this line to overflow as the PDF compile will not break the formatted code into separate lines.  Please find another way to write this.  For example, you could state "Polytopes computed by `constrained_manipulability` are published using ROS messages from the `visualization_msgs::msg` namespace. The primary type is `MarkerArray`, which utilizes `Marker::TRIANGLE_LIST` markers for facets and `Marker::SPHERE_LIST` markers for vertices."
- [ ] LINE 101:  The citation "...of (Zolotas et al., 2021)." Should appear as "...of Zolotas et al (2021)."  You can do this in your paper.md file using "Eq. 20 of @Zolotas2021Motion." instead of "Eq. 20 of [@Zolotas2021Motion]."
- [ ] LINE 106:  Same as the above comment, should be "Long et al. (2019)."

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.
