+++ paper/paper.md
@@ -42,7 +42,7 @@ The library is encapsulated in a Robot Operating System (ROS) package [@Quigley2
 
 The main program of the `constrained_manipulability` package reads a Unified Robot Description Format (URDF) kinematic chain from a user-defined root of the robot (e.g., the robot's base frame) to a tip (e.g., the end-effector or tool). Joint position and velocity limits are also read from the URDF, while a collision world is maintained to express environmental constraints on the robot's motion. A variety of collision objects can be added or removed using common ROS message types, such as OctoMaps [@Hornung2013Octomap], mesh files, or solid primitives. These joint limit and obstacle constraints are then stacked together to define different polytopes, such as the allowable motion polytope [@Long2019Optimization] and the constrained velocity polytope [@Long2018Humanoids].
 
-Polytopes computed by `constrained_manipulability` are published as a `visualization_msgs::msg::MarkerArray` ROS message. A `visualization_msgs::msg::Marker::TRIANGLE_LIST` represents the polytope facets and a `visualization_msgs::msg::Marker::SPHERE_LIST` represents the vertices. As a result, the manipulability polytopes  calculated by our package for a given robot manipulator can also be visualized using standard RViz [@kam2015rviz] tools. Additionally, the generated polytopes are published in vertex and hyperplane form, which is easily interpretable by third-party libraries.
+Polytopes computed by `constrained_manipulability` are published using ROS messages from the `visualization_msgs::msg` namespace. The primary type is `MarkerArray`, which utilizes `Marker::TRIANGLE_LIST` markers for facets and `Marker::SPHERE_LIST` markers for vertices. As a result, the manipulability polytopes  calculated by our package for a given robot manipulator can also be visualized using standard RViz [@kam2015rviz] tools. Additionally, the generated polytopes are published in vertex and hyperplane form, which is easily interpretable by third-party libraries.
 
 # Statement of Need
 
@@ -91,9 +91,9 @@ where $\dot{\mathbf{q}}^{v}_{i}$ denotes the $i^{th}$ vertex of the polytope $\m
 \end{equation}
 with $\boldsymbol{\nu}^{v}_{j} = \mathbf{J}_{n} \dot{\mathbf{q}}^{v}_{j}$ and $p \leq q$. The convexity of a polytope is preserved under affine transformation, thus a bounded volume of $\mathcal{MP}$ that represents the system's manipulability can easily be obtained to serve as an exact indicator of robot performance. 
 
-For constrained motion polytopes, the following constraints are supported: joint velocity limits, joint velocity damping due to positional joint limits, and joint velocity damping due to obstacle proximity to the kinematic chain. For allowable motion polytopes the following constraints are supported: positional joint limits, positional limits due to obstacle proximity to the kinematic chain, and linearization limits. Increasing the values of the linearization expands the free space virtual fixture at a cost of reduced fidelity. Our implementation uses a scalar linearization limit for all joints that can also be altered at run time, e.g., could be increased to enlarge the solution or shrunk to guide the user towards a defined goal configuration, as defined in Eq. 20 of [@Zolotas2021Motion]. Finally, in our implementation, we select at every instant the point along each link nearest to all environmental obstacles. Hence, the system considers the set of instantaneous collision-free joint motion for each link along the robot's kinematic chain, considering all surrounding obstacles. 
[…]
