### 2025-04-22 #7481 @mazrk7
Hi @crvernon, I've made the changes according to your suggestions.

The DOI of the latest archived version on Zenodo is: [10.5281/zenodo.15258944](https://doi.org/10.5281/zenodo.15258944)

And the new version number released is: v1.1

Please let me know if that all looks good, thanks.
